refactor: ask 관련 파일 경로 수정 (#36)
